+++ src/main/java/model/managers/IBasketManager.java
@@ -3,9 +3,9 @@ package model.managers;
 import model.utils.Size;
 
 public interface IBasketManager {
-    void addPizzaToBasket(long userId, String pizzaName, Size size);
-    void deletePizzaFromBasket(long userId, String pizzaName, Size size);
+    void addPizzaToBasket(int userId, String pizzaName, Size size);
+    void deletePizzaFromBasket(int userId, String pizzaName, Size size);
 
     //returns orderId
-    long createOrder(long userId, String userAddress, String userPhone);
+    long createOrder(int userId, String userAddress, String userPhone);
 }
